+++ src/library/inputs/xcbpointer.h
@@ -1,0 +1,95 @@
+/*
+    Copyright 2015-2018 Clément Gallet <clement.gallet@ens-lyon.org>
+
+    This file is part of libTAS.
+
+    libTAS is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    libTAS is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with libTAS.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef LIBTAS_XCBKEYBOARD_H_INCL
+#define LIBTAS_XCBKEYBOARD_H_INCL
+
+#include "../global.h"
+#include <xcb/xcb.h>
+#include <xcb/xcb_keysyms.h>
+#include <xcb/xproto.h>
+
+namespace libtas {
+
+OVERRIDE xcb_query_pointer_cookie_t xcb_query_pointer (xcb_connection_t *c, xcb_window_t window);
+
+OVERRIDE xcb_query_pointer_cookie_t xcb_query_pointer_unchecked (xcb_connection_t *c, xcb_window_t window);
+
+OVERRIDE xcb_query_pointer_reply_t *xcb_query_pointer_reply (xcb_connection_t *c, xcb_query_pointer_cookie_t cookie, xcb_generic_error_t **e);
+
+OVERRIDE xcb_void_cookie_t
+xcb_warp_pointer_checked (xcb_connection_t *c,
+                          xcb_window_t      src_window,
+                          xcb_window_t      dst_window,
+                          int16_t           src_x,
+                          int16_t           src_y,
+                          uint16_t          src_width,
+                          uint16_t          src_height,
+                          int16_t           dst_x,
+                          int16_t           dst_y);
+
+OVERRIDE xcb_void_cookie_t
+xcb_warp_pointer (xcb_connection_t *c,
+                  xcb_window_t      src_window,
+                  xcb_window_t      dst_window,
+                  int16_t           src_x,
+                  int16_t           src_y,
+                  uint16_t          src_width,
+                  uint16_t          src_height,
+                  int16_t           dst_x,
+                  int16_t           dst_y);
+
+OVERRIDE xcb_grab_pointer_cookie_t
+xcb_grab_pointer (xcb_connection_t *c,
+                  uint8_t           owner_events,
+                  xcb_window_t      grab_window,
+                  uint16_t          event_mask,
+                  uint8_t           pointer_mode,
+                  uint8_t           keyboard_mode,
+                  xcb_window_t      confine_to,
+                  xcb_cursor_t      cursor,
+                  xcb_timestamp_t   time);
+
+OVERRIDE xcb_grab_pointer_cookie_t
+xcb_grab_pointer_unchecked (xcb_connection_t *c,
+                            uint8_t           owner_events,
+                            xcb_window_t      grab_window,
+                            uint16_t          event_mask,
+                            uint8_t           pointer_mode,
+                            uint8_t           keyboard_mode,
+                            xcb_window_t      confine_to,
+                            xcb_cursor_t      cursor,
+                            xcb_timestamp_t   time);
+
+OVERRIDE xcb_grab_pointer_reply_t *
+xcb_grab_pointer_reply (xcb_connection_t           *c,
+                        xcb_grab_pointer_cookie_t   cookie  /**< */,
+                        xcb_generic_error_t       **e);
+
+OVERRIDE xcb_void_cookie_t
+xcb_ungrab_pointer_checked (xcb_connection_t *c,
+                            xcb_timestamp_t   time);
+
+OVERRIDE xcb_void_cookie_t
+xcb_ungrab_pointer (xcb_connection_t *c,
+                    xcb_timestamp_t   time);
+
+}
+
+#endif